# OpenAPISpecification.yaml
swagger: '2.0'
info:
  description: ' REST APIs for Gowma Online Store'
  version: '1.0'
  title: GOWMA service API
  termsOfService: Terms of service
  contact:
    name: Gagan Deep
    url: 'https://org.mihy.gowma/about/'
    email: gagandeepts@gmail.com
  license:
    name: Apache License Version 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8081'
basePath: /gowma-service
...
paths:
...
  /api/v1/user:
    post:
      tags:
        - user-controller
      summary: Create a user
      operationId: createUserUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: user
          description: user
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/v1/user/{user-id}':
    get:
      tags:
        - user-controller
      summary: Get a user by id
      operationId: getUserByIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - user-controller
      summary: Update a user by id
      operationId: updateUserUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
        - in: body
          name: user
          description: user
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - user-controller
      summary: Delete a user by id
      operationId: deleteUserUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/api/v1/user/{user-id}/addresses':
    get:
      tags:
        - user-controller
      summary: Get a list of addresses by a user id
      operationId: getAddressesForUserIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserAddress'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - user-controller
      summary: Add a new address to the user by id
      operationId: createAddressForUserIdUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
        - in: body
          name: userAddress
          description: userAddress
          required: true
          schema:
            $ref: '#/definitions/UserAddress'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/UserAddress'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/v1/user/{user-id}/addresses {address-id}':
    post:
      tags:
        - user-controller
      summary: Update a existing address of user by id
      operationId: updateAddressForUserIdUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
        - name: address-id
          in: path
          description: address-id
          required: true
          type: integer
          format: int32
        - in: body
          name: userAddress
          description: userAddress
          required: true
          schema:
            $ref: '#/definitions/UserAddress'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/UserAddress'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - user-controller
      summary: 'Delete a address of a user by id '
      operationId: deleteAddressForUserIdUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
        - name: address-id
          in: path
          description: address-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/api/v1/user/{user-id}/cart':
    get:
      tags:
        - user-controller
      summary: Get a list of shopping cart items for a user by user id
      operationId: getShoppingCartForUserIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ShoppingCartItem'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - user-controller
      summary: Add a shopping cart item to the user
      operationId: addToShoppingCartUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
        - in: body
          name: shoppingCartItem
          description: shoppingCartItem
          required: true
          schema:
            $ref: '#/definitions/ShoppingCartItem'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ShoppingCartItem'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - user-controller
      summary: Clear the whole cart list for user
      operationId: clearAllShoppingCartListForUserUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/api/v1/user/{user-id}/product-wish-list':
    get:
      tags:
        - user-controller
      summary: 'Get user wishlist by user id '
      operationId: getWishedProductsForUserIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/UserWishedProduct'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - user-controller
      summary: Clear the whole user wishList
      operationId: clearUserWishListUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/api/v1/user/{user-id}/product-wish-list/{product-id}':
    post:
      tags:
        - user-controller
      summary: Add a product with given id to the user
      operationId: addProductToWishListUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
      responses:
        '201':
          description: Created
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - user-controller
      summary: Remove a product with given id from the user
      operationId: removeProductToWishListUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user-id
          in: path
          description: user-id
          required: true
          type: integer
          format: int32
        - name: product-id
          in: path
          description: product-id
          required: true
          type: integer
          format: int32
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
definitions:
  AddressCode:
    type: object
    properties:
      code:
        type: string
      countryName:
        type: string
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      placeName:
        type: string
      stateName:
        type: string
...
  Product:
    type: object
    properties:
      active:
        type: boolean
      caption:
        type: string
      categoryId:
        type: integer
        format: int32
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      name:
        type: string
      price:
        type: number
        format: double
      unitOfMeasure:
        $ref: '#/definitions/UnitOfMeasure'
      viewCount:
        type: integer
        format: int32
...
  Role:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      name:
        type: string
  ShoppingCartItem:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      product:
        $ref: '#/definitions/Product'
      quantity:
        type: number
        format: double
      userId:
        type: integer
        format: int32
...
  UnitOfMeasure:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      name:
        type: string
  User:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      email:
        type: string
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      mobileNo:
        type: string
      password:
        type: string
      roles:
        type: array
        items:
          $ref: '#/definitions/Role'
      status:
        type: string
        enum:
          - ACTIVE
          - INACTIVE
      userDetail:
        $ref: '#/definitions/UserDetail'
  UserAddress:
    type: object
    properties:
      address1:
        type: string
      address2:
        type: string
      addressCode:
        $ref: '#/definitions/AddressCode'
      addressName:
        type: string
      addressType:
        type: string
        enum:
          - OFFICE
          - HOME
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      default:
        type: boolean
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      landmark:
        type: string
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      phoneNo:
        type: string
      userId:
        type: integer
        format: int32
  UserDetail:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      fname:
        type: string
      gender:
        type: string
        enum:
          - MALE
          - FEMALE
          - OTHER
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      lname:
        type: string
      userId:
        type: integer
        format: int32
  UserWishedProduct:
    type: object
    properties:
      createdBy:
        type: integer
        format: int32
      createdDate:
        type: string
        format: date-time
      deleted:
        type: boolean
      id:
        type: integer
        format: int32
      lastModifiedBy:
        type: integer
        format: int32
      lastModifiedDate:
        type: string
        format: date-time
      product:
        $ref: '#/definitions/Product'
      userId:
        type: integer
        format: int32